Close connection after a certain timeout
If the client has established a connection, but then idles for 5
minutes, the connection is aborted by the server. Any data from the
client will postpone the timeout by 5 minutes (like a debounce).

RFC 2821 section 4.5.3.2 suggests different timeouts after different
commands, but we're keeping it simple with 5 minutes for all.
https://www.rfc-editor.org/rfc/rfc2821#section-4.5.3.2

Not sure if it's customary to send 221 to the client before closing, but
I guess that's a nice thing to do.